refactor (new-product): removing equals and hashCode from Form
